Move FeatureLauncher to dedicated folder
