+++ resources/assets/components/Campaign/Campaign.js
@@ -5,8 +5,8 @@ import ReactRouterPropTypes from 'react-router-prop-types';
 import Modal from '../utilities/Modal/Modal';
 import NotificationContainer from '../Notification';
 import ModalRoute from '../utilities/ModalRoute/ModalRoute';
-import FeatureLauncher from '../utilities/Modal/ModalLauncher';
 import TypeFormSurvey from '../utilities/TypeFormSurvey/TypeFormSurvey';
+import FeatureLauncher from '../utilities/FeatureLauncher/FeatureLauncher';
 import CampaignRouteContainer from './CampaignRoute/CampaignRouteContainer';
 import TrafficDistribution from '../utilities/TrafficDistribution/TrafficDistribution';
 import VoterRegistrationModal from '../pages/VoterRegistrationModal/VoterRegistrationModal';
